+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.8
+
+### Patch Changes
+
+- Support for optional parameters
+- Updated dependencies
+  - @beff/cli@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes